Tested E-Mail Receiving

+++ src/main/java/htlle/mailresponse/Repository/EmailDummyRepository.java
@@ -3,6 +3,6 @@ package htlle.mailresponse.Repository;
 import htlle.mailresponse.Mail.EmailDummy;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EmailDummyRepository extends JpaRepository<EmailDummy, Long> {
+public interface EmailDummyRepository extends JpaRepository<EmailDummy, Integer> {
     EmailDummy findById(int id);
 }
